Added basic ToJSON typeclass
