+++ stack.yaml
@@ -1,1 +1,7 @@
-resolver: https://raw.githubusercontent.com/nikita-volkov/stack-snapshot/4163770ee93e3c3813a7289c843565d1b9ab98f0/snapshot.yaml
+#resolver: https://raw.githubusercontent.com/nikita-volkov/stack-snapshot/4163770ee93e3c3813a7289c843565d1b9ab98f0/snapshot.yaml
+resolver: lts-16.18
+
+extra-deps:
+- {git: "https://github.com/nikita-volkov/ptr-poker", commit: "75176794a789883da559840e32ea85ee345035ec"}
+- numeric-limits-0.1.0.0
+- unordered-containers-0.2.13.0
